Fixing tests, breaking circular dependencies
