+++ kifi-backend/modules/sqldb/app/com/keepit/common/db/slick/Database.scala
@@ -158,11 +158,6 @@ class Database @Inject() (
         }
       } finally { rw.close() }
     }
-    if (failure != null) {
-      //this is very dangerous since we swallow the exception (just report it) and don't let the caller know something bad happened as an exception
-      //the signature of the method should be revisited to enforce check of partial failure of the transaction
-      airbrake.notify(s"Batch partial fail (${failure.exception.getClass.getSimpleName}) readWrite transaction, processed $successCnt out of ${batch.size}", failure.exception)
-    }
     results
   }
